[Lucas Yonghee Lee] renewal/6 : fix #FG-WEBADMIN-API-RENEWAL/6
Modify /ad/getcategoryaddetail API

# Conflicts:
#	src/main/java/net/fashiongo/webadmin/service/renewal/RenewalAdService.java

+++ src/main/java/net/fashiongo/webadmin/data/entity/primary/AdVendorEntity.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.Set;
 
@@ -39,7 +40,7 @@ public class AdVendorEntity {
 	private Integer vendorCategoryID;
 
 	@Column(name = "ActualPrice")
-	private Integer actualPrice;
+	private BigDecimal actualPrice;
 
 	@Column(name = "FromDate")
 	private Timestamp fromDate;
